chore: move WebSocketClassOverride into separate module
